chore: rename ws to app
