+++ apps/server/src/event-handlers/takeout.ts
@@ -1,6 +1,6 @@
 import type { CoreEvent, CoreEventData } from '@tg-search/core'
+import type { ClientState } from '../app'
 import type { WsMessageToServer } from '../utils/ws-event'
-import type { ClientState } from '../ws'
 
 import { useLogger } from '@tg-search/common'
